backend/requirements.txt
# FastAPI + servidor
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0

# Base de datos
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Validación y configuración
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# Autenticación
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.20

# Tareas asíncronas (Fase 6)
# celery==5.4.0
# redis==5.2.1

# Utilidades
python-dateutil==2.9.0
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
